refactor: createShouldForwardProp from createX
